feat: add user login/logout. Add templates and refactor. Add user pro…

+++ core/views.py
@@ -1,13 +1,62 @@
+from django.contrib import auth
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Messages
 from .forms import RoomForm
 
-# rooms = [
-#     {'id': 1, 'name': 'room 1'},
-#     {'id': 2, 'name': 'room 2'},
-#     {'id': 3, 'name': 'room 3'},
-# ]
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'User or Password does not exists')
+
+    context = {'page': page}
+    return render(request, 'core/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error ocurred during registration')
+    return render(request, 'core/login_register.html', {'form': form})
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -15,19 +64,53 @@ def home(request):
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
         Q(description__icontains=q)
-        )
+    )
 
     topics = Topic.objects.all()
     room_count = rooms.count()
+    room_messages = Messages.objects.filter(Q(room__topic__name__icontains=q))  # filter comments by topic
 
-    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
+    context = {
+        'rooms': rooms, 
+        'topics': topics, 
+        'room_count': room_count,
+        'room_messages': room_messages
+        }
     return render(request, 'core/home.html', context)
+
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.messages_set.all()
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Messages.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {'room': room, 'room_messages': room_messages,
+               'participants': participants}
     return render(request, 'core/room.html', context)
 
+def userProfile(request, pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.messages_set.all()
+    topics = Topic.objects.all()
+    context = {
+        'user': user, 
+        'rooms': rooms, 
+        'room_message': room_messages,
+        'topics': topics
+        }
+    return render(request, 'core/profile.html', context)
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -38,9 +121,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'core/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -51,10 +139,27 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'core/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'core/delete.html', {'obj': room})
 
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Messages.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here.')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'core/delete.html', {'obj': message})
